Fix Plasmoid/Applet API docs
This addresses the issues mentioned in https://invent.kde.org/plasma/libplasma/-/merge_requests/1374 and a few more.

For now, just use Plasmoid for the QML API and Applet for the C++ API because it would be weird otherwise. We can change this later once we know what to do with the naming.